refactor(adms): split client.py by entity for navigability
Per reviewer feedback (NicoleMGomes, PR #138): client.py was 1432 lines
holding 8 internal API classes plus the public clients/factories. Move
the 8 classes into 4 sibling modules paired by entity (sync + async
together to prevent drift):

- _document_api.py             — _DocumentApi + _AsyncDocumentApi
- _relation_api.py             — _DocumentRelationApi + _AsyncDocumentRelationApi
- _configuration_api.py        — _ConfigurationApi + _AsyncConfigurationApi
- _job_api.py                  — _JobApi + _AsyncJobApi

client.py drops to 217 lines holding only AdmsClient, AsyncAdmsClient,
create_client, create_async_client. Tests that import the private
classes via `sap_cloud_sdk.adms.client` keep working — the new client.py
imports them at the top level so the original import paths still resolve.

Public surface (everything in adms/__init__.py's __all__) is unchanged.
